+++ CHANGELOG.md
@@ -4,15 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [5.1.0] - 2022/01/21
 
-## Fixed
+## Added
 
-* Summary urls are build up correctly
+* Blocked test cases comments now also get a badge
 
 ## Changed
 
 * The date format in comments is formatted now as %Y-%m-%d %H:%M
+
+## Fixed
+
+* Summary urls are build up correctly
+* Prevent the same comment in testcases
 
 ## [5.0.1] - 2021/11/30